+++ packages/band/src/index.tsx
@@ -1,4 +1,5 @@
 export * from './bands'
 export * from './bars'
 export * from './quantiles'
+export * from './schedules'
 export * from './strips'