+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 
 ### Added
 
--   [core]: support for time scales and Date objects
+-   [packages/band]: components `Schedule` and `Schedules` for schedule / Gantt charts
+-   [examples]: examples of line charts with time scales
+-   [packages/core]: support for time scales
 
+### Changed
+
+-   [documentation]: miscellaneous passages in documentation pages
 
 ## v0.1.0